pyproject.toml
[project]
name = "joi"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.2",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "ty>=0.0.13",
    "aiogram>=3.24.0",
    "openai>=2.15.0",
    "fastmcp>=2.14.4",
    "tmdbsimple>=2.9.1",
    "uvicorn>=0.40.0",
    "fastapi>=0.128.0",
    "telegramify-markdown>=0.5.4",
    "loguru>=0.7.3",
    "sqlalchemy[asyncio]>=2.1.0b1",
    "transmission-rpc>=8.0.0a4",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "psycopg[binary]>=3.3.2",
    "jmespath>=1.1.0",
    "httpx>=0.28.1",
    "xmltodict>=1.0.2",
    "langgraph>=1.0.7",
    "langchain-openai>=1.1.7",
    "langchain-mcp-adapters>=0.2.1",
    "langfuse>=3.12.1",
    "langchain>=1.2.8",
    "langgraph-cli[inmem]>=0.4.12",
    "langmem>=0.0.30",
    "mem0ai>=0.1.115",
    "langchain-anthropic>=1.3.1",
    "pydantic-monty>=0.0.4",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "pytest-recording>=0.13.4",
    "ruff>=0.14.0",
    "watchfiles>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/joi_agent_langgraph2", "src/joi_mcp", "src/joi_telegram_langgraph", "src/joi_langgraph_client"]

[tool.uv]
prerelease = "allow"

[tool.ruff]
target-version = "py312"
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ty.environment]
python-version = "3.12"
python = "./.venv"

[tool.ty.src]
include = ["tests"]

[tool.ty.rules]
# Core type safety - error
possibly-unresolved-reference = "error"
invalid-argument-type = "error"
missing-argument = "error"
unsupported-operator = "error"
division-by-zero = "error"

# Disabled — false positives from third-party stubs (ChatOpenAI kwargs, FunctionTool.__call__)
unknown-argument = "ignore"
call-non-callable = "ignore"

# Code quality - warn
unused-ignore-comment = "warn"
redundant-cast = "warn"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
pythonpath = "."
markers = [
    "contract: marks tests as contract tests (external API verification)",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "eval: marks tests as LLM evaluation tests (hits real API)",
    "e2e: marks tests as end-to-end tests (requires running services)",
]
